api/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
pillow
opencv-python-headless
pydantic
gunicorn
pytest
httpx
